+++ packages/metastable/src/sysinfo/gpu/nvidia.ts
@@ -1,58 +1,67 @@
 import fs from 'fs/promises';
+import os from 'os';
 import path from 'path';
 
+import which from 'which';
+
 import { parseNumber } from '../../helpers/common.js';
-import { shell } from '../../helpers/spawn.js';
-import { GraphicsControllerData } from '../types.js';
+import { stdout } from '../../helpers/spawn.js';
+import {
+  GraphicsControllerData,
+  GraphicsControllerUtilization,
+} from '../types.js';
 import * as util from '../util.js';
 
-const _platform = process.platform;
 let _nvidiaSmiPath: string | undefined = undefined;
 let _nvidiaSmiRetry = true;
 
-const _linux = _platform === 'linux' || _platform === 'android';
-const _windows = _platform === 'win32';
-
 async function getNvidiaSmi() {
-  if (!_nvidiaSmiRetry) {
-    return false;
-  }
-
   if (_nvidiaSmiPath) {
     return _nvidiaSmiPath;
   }
 
-  if (_windows) {
-    try {
-      const basePath = util.WINDIR + '\\System32\\DriverStore\\FileRepository';
-      // find all directories that have an nvidia-smi.exe file
-      const candidates = await Promise.all(
-        (await fs.readdir(basePath)).map(async dir => {
-          const filePath = path.join(basePath, dir, 'nvidia-smi.exe');
-          try {
-            const stat = await fs.stat(filePath);
-            return {
-              filePath,
-              stat,
-            };
-          } catch {
-            return undefined;
-          }
-        }),
-      );
-      const sorted = candidates
-        .filter(item => typeof item === 'object')
-        .sort((a, b) => b.stat.ctimeMs - a.stat.ctimeMs);
-      const first = sorted[0].filePath;
-      if (first) {
-        _nvidiaSmiPath = first;
+  if (!_nvidiaSmiRetry) {
+    return false;
+  }
+
+  switch (os.platform()) {
+    case 'win32':
+      try {
+        const basePath =
+          util.WINDIR + '\\System32\\DriverStore\\FileRepository';
+        // find all directories that have an nvidia-smi.exe file
+        const candidates = await Promise.all(
+          (await fs.readdir(basePath)).map(async dir => {
+            const filePath = path.join(basePath, dir, 'nvidia-smi.exe');
+            try {
+              const stat = await fs.stat(filePath);
+              return {
+                filePath,
+                stat,
+              };
+            } catch {
+              return undefined;
+            }
+          }),
+        );
+        const sorted = candidates
+          .filter(item => typeof item === 'object')
+          .sort((a, b) => b.stat.ctimeMs - a.stat.ctimeMs);
+        const first = sorted[0].filePath;
+        if (first) {
+          _nvidiaSmiPath = first;
+        }
+      } catch {
+        _nvidiaSmiRetry = false;
       }
-    } catch {
-      _nvidiaSmiRetry = false;
-    }
-  } else if (_linux) {
-    _nvidiaSmiPath = 'nvidia-smi';
-    // TODO: Check if present.
+      break;
+    case 'linux':
+      try {
+        _nvidiaSmiPath = await which('nvidia-smi');
+      } catch {
+        _nvidiaSmiRetry = false;
+      }
+      break;
   }
 
   return _nvidiaSmiPath;
@@ -62,16 +71,17 @@ export async function nvidiaSmi<
   TColumns extends string[],
   TResult = { [K in TColumns[number]]: string | undefined },
 >(columns: readonly [...TColumns]): Promise<TResult[]> {
-  const nvidiaSmiExe = await getNvidiaSmi();
-  if (!nvidiaSmiExe) {
+  const nvidiaSmiPath = await getNvidiaSmi();
+  if (!nvidiaSmiPath) {
     throw new Error('Unable to find nvidia-smi');
   }
 
-  const nvidiaSmiOpts = `--query-gpu=${columns.join(',')} --format=csv,noheader,nounits`;
-  const cmd =
-    nvidiaSmiExe + ' ' + nvidiaSmiOpts + (_linux ? '  2>/dev/null' : '');
-
-  const output = (await shell(cmd)).split('\n');
+  const output = (
+    await stdout(nvidiaSmiPath, [
+      `--query-gpu=${columns.join(',')}`,
+      '--format=csv,noheader,nounits',
+    ])
+  ).split('\n');
   const result: any[] = [];
   for (const line of output) {
     const split = line.split(', ');
@@ -126,19 +136,12 @@ export async function nvidiaDevices(): Promise<GraphicsControllerData[]> {
     return {
       subDeviceId: gpu['pci.sub_device_id'],
       name: gpu['name'],
+      vendor: 'NVIDIA',
       pciBus: gpu['pci.bus_id'],
-      fanSpeed: parseNumber(gpu['fan.speed']),
       memoryTotal: vram,
       memoryUsed: parseMemory(gpu['memory.used']),
-      memoryFree: parseMemory(gpu['memory.free']),
       utilizationGpu: parseNumber(gpu['utilization.gpu']),
-      utilizationMemory: parseNumber(gpu['utilization.memory']),
       temperatureGpu: parseNumber(gpu['temperature.gpu']),
-      temperatureMemory: parseNumber(gpu['temperature.memory']),
-      powerDraw: parseNumber(gpu['power.draw']),
-      powerLimit: parseNumber(gpu['power.limit']),
-      clockCore: parseNumber(gpu['clocks.gr']),
-      clockMemory: parseNumber(gpu['clocks.mem']),
       vram: vram,
       vramDynamic: false,
       bus: 'PCI',
@@ -154,13 +157,14 @@ const utilizationColumns = [
   'temperature.gpu',
 ] as const;
 
-export async function nvidiaUtilization() {
+export async function nvidiaUtilization(): Promise<
+  GraphicsControllerUtilization[]
+> {
   const gpus = await nvidiaSmi(utilizationColumns);
 
   return gpus.map(gpu => ({
     memoryTotal: parseMemory(gpu['memory.total']),
     memoryUsed: parseMemory(gpu['memory.used']),
-    memoryFree: parseMemory(gpu['memory.free']),
     utilizationGpu: parseNumber(gpu['utilization.gpu']),
     temperatureGpu: parseNumber(gpu['temperature.gpu']),
   }));